Propose a class-based input system
Rewritten model specification through a dictionary referred to as spec.
Two functions for computing det parameter. I am differentiating between
model constants that define the model - these are in encapsulated in
spec and all the variables necessary to run the model - these have been
moved to class ModelInput. Also, the RHS of the differential equatio is
also a class (functor) to avoid long list of arguments.