Corrected indexing on line 128 of preprocessing,py
The indexing on this line hadn't been updated in the translation to Python to account for zero-indexing. In the Matlab version we are assigning the first system_sizes[1] rows, so here we need to assign the first system_sizes[0] rows. The incorrect indexing happens to not cause any problems in the UK-like example, but causes an error if, for example, Kenya-like demographic data with three age classes is used.