Moved excel and cvs reads to specs file
All of the read_excel and read_csv operations from model.preprocessing are now done in model.specs. Moving the read operations in this way should make it easier to adapt the code to other population settings. Moving forward, it might make sense to store each national setting in a separate dictionary in model.specs.