Added exponentially growing imports
The class RateEquationsWithExponentialImports takes "profiles" of detected and undetected cases as input - these should be the sections of the early growth eigenvector corresponding to these cases, plus the exponential growth rate r. In the working example, r is taken to be gamma*R0-gamma, and the profiles are just the detection/undetection probabilities used in the example, although it would be more realistic to work out the next generation matrix and use its eigenvalue/eigenvector.