Began implementing likelihood calculations from Vo data
run_vo_households.py should calculate the probability of obtaining the Vo testing data given the parameters in specs.py and preprocessing.py. It doesn't currently run, but the code as it appears should give an outline of how the calculation will be done. The main challenge here is adapting the existing code to allow for a single household type, which is not currently allowed.