Vo example running but giving bad output
The Vo example is now running, but is clearly getting something wrong because it calculates the likelihood as NaN. We possibly need to do a log likelihood sum instead of a product. We skip the three largest households (sizes 11, 24, and 44) due to memory and time constraints. A function get_symptoms_by_test has been added to model\common which returns which of a list of test results where symptomatic and asymptomatic. The within_household_mixing in the Vo example has been replaced with build_household_population applied to a single-type household type list, which appears to work.