Fixed divide by zero error in get_FOI_by_class
There were problems with get_FOI_by_class because if not all classes are present in the set of compositions, you get a divide by zero. This has now been corrected for the exponential imports version but not for any of the others. The Vo example is still giving bad (not in [0,1]) probablities.