Added Vo example to master
This version, unlike the one in the household-testing-likelihood-calculations branch, is consistent with the more modularised approach to the rate equations. I have slightly modified the det_by_class and undet_by_class calculations in get_FOI_by_class to allow for composition lists where not every age class is represented - this is needed for doing single-household populations in the Vo example.