Create external isolation branch
A new function, within_household-spread_with_isolation has been added to common. Ths function constructs the transition matrix for within-household events with an "isolated" class. The isolation rate is currently the same for all age classes but can be different for detected and undetected cases. The age classes which are allowed to isolated can be chosen by the user and for households with children, at least one adult must be at home not isolating at any one time. The external_isolation example currently just contains an identical copy of external_imports\run.py but this will be updated to give a working example of the external isolation policy.