Introduced state matrix construction
The fuctions build_state_matrix build_states_recursively and state_recursor act to recursively construct the states object. This means that to build any given system we just need to call build_state_matrix rather than doing a big nested for loop in the construction itself. Since this for loop just depends on the number of compartments and is independent of what the compartments actually mean, having this set of functions means we avoid writing the same set of loops every time we look at a new system.