Create parallel_sweep_vacc_efficacy.py
This new parallel sweep looks at the effect of different levels of efficacy when we have high patient and staff uptake. The intention is to have an output which tells you where the transition from subcritical to supercritical infection takes place, although it doesn't quite seem to be doing this effectively yet.