Fixed scale problem in temp_bubbles example
The problem was arising because I had changed the way I assembled the index_vector list. This has now been corrected, and we get generally similar results to before Christmas, although with much more infection, probably due to a parameter change. There may be numerical errors after implementing the merges which mean you see a sharp jump in infection, but this could also just be a very fast period of growth emerging when we redefine the household structure.