Began reformatting OOHI example
This is only partly done - I am pushing these changes to master so that I can switch branches to work on parallelising the temporary bubbling example. I am part way through developing an `add_vuln_class` function, which will convert the model input from the 2-class model to one with adults divided into more and less vulnerable categories. I have tried to do this in a very generic way so that it can be adapted to models with more or fewer age classes.