Confirmed that parallelised and serial sweeps give the same results
The parallelised version of the temporary bubbles parameter sweep executes, and gives the same results as the serial version (currently on the master branch). The parallel_bubbles branch can now be merged with the master branch. This example now saves results to an output folder rather than the base directory; going forward I will be editing all the examples to do this.