Made growth rate calculation less model-specific
The growth rate was coded specifically to work for the SEPIR framework, but not the SIR or SEIR models. I have made it more general by adding a "new case compartment" to the model input, which allows the growth rate calculator to find the index states for  general compartmental structure.