Added sketch of initial condition to single type construction
Note: this currently only includes a sketch for an initial condition of type (N-i, 0, i, 0), i.e. assuming there are no incubating cases or recovered individuals in the household. In general the initial condition without this restriction is nontrivial to calculate since it depends on the household's past exposure to infection.