Added capability for specifying internal transmission
In the definitions for the different model inputs, there is now a check to see if a value for beta_int and density_expo have been specified in the spec dictionary. If a dictionary key for beta_int exists, the corresponding value is used in the model input. If one exists for density_expo, then the corresponding value is used in the model input, otherwise the density exponent is set to 1 (frequency-dependent mixing). If there is no dictionary key for beta_int, the init statement looks for an SITP in the dictionary and uses least squares fitting to estimate beta_int and the density exponent.