Method Comparison
Synthetic data generator generates a pickle file using the data generating function we already had.

Likelihood method inference is the old version of the mle calculation but I have added some code so it can also print the loglikelihood.

gradient method inference, is the new method, calculating the gradients. I have expanded a little on our latest work and now it calculates the gradients correctly and, also calculates the llikelihood of all households in the data.

The simulation inference file now does not include what is in in the gradient method file for clarity.